# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API REST Hoteles",
        "description": "API REST Hoteles",
        "contact": {
            "email": "cesarpozo25@gmail.com"
        },
        "version": "1.0.0"
    },
    "host": "localhost/apirest-hoteles/public/",
    "basePath": "",
    "paths": {
        "/alojamientos": {
            "get": {
                "summary": "Get Alojamientos",
                "parameters": [
                    {
                        "name": "Autorization",
                        "in": "header",
                        "description": "Bearer {token_access}",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#definitions/Alojamiento"
                            }
                        }
                    },
                    "403": {
                        "description": "Autorization Exception",
                        "schema": {
                            "$ref": "#definitions/Errors403"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#definitions/Errors500"
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
        "/alojamientos/{alojamiento_id}": {
            "get": {
                "summary": "Show one Alojamiento",
                "parameters": [
                    {
                        "name": "Autorization",
                        "in": "header",
                        "description": "Bearer {token_access}",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "alojamiento_id",
                        "in": "path",
                        "description": "un numero id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#definitions/Alojamiento"
                            }
                        }
                    },
                    "403": {
                        "description": "Autorization Exception",
                        "schema": {
                            "$ref": "#definitions/Errors403"
                        }
                    },
                    "404": {
                        "description": "Not Found Exception",
                        "schema": {
                            "$ref": "#definitions/Errors404"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#definitions/Errors500"
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "put": {
                "summary": "Update Alojamientos",
                "parameters": [
                    {
                        "name": "Autorization",
                        "in": "header",
                        "description": "Bearer {token_access}",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "alojamiento_id",
                        "in": "path",
                        "description": "un numero id",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "properties": {
                                "precio": {
                                    "type": "string",
                                    "example": "299.99"
                                }
                            }
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#definitions/Alojamiento"
                            }
                        }
                    },
                    "403": {
                        "description": "Autorization Exception",
                        "schema": {
                            "$ref": "#definitions/Errors403"
                        }
                    },
                    "404": {
                        "description": "Not Found Exception",
                        "schema": {
                            "$ref": "#definitions/Errors404"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#definitions/Errors500"
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            },
            "delete": {
                "summary": "Delete Alojamientos",
                "parameters": [
                    {
                        "name": "Autorization",
                        "in": "header",
                        "description": "Bearer {token_access}",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "alojamiento_id",
                        "in": "path",
                        "description": "un numero id",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#definitions/Alojamiento"
                            }
                        }
                    },
                    "403": {
                        "description": "Autorization Exception",
                        "schema": {
                            "$ref": "#definitions/Errors403"
                        }
                    },
                    "404": {
                        "description": "Not Found Exception",
                        "schema": {
                            "$ref": "#definitions/Errors404"
                        }
                    },
                    "500": {
                        "description": "internal server error",
                        "schema": {
                            "$ref": "#definitions/Errors500"
                        }
                    }
                },
                "security": [
                    {
                        "passport": []
                    }
                ]
            }
        },
...
    },
...
}